Split server.rs into ten sub-modules grouped by surface (F-12)
server.rs had grown to 2000 LOC holding every route, every payload struct, response helpers, the pty/runner WebSocket handlers, path validators, and the log_non_2xx middleware. One file per surface makes the route map navigable and lets each handler group declare only the imports it actually needs.

Layout (all files under src-tauri/src/server/):

- mod.rs — doc header, AppState, start, build_router, log_non_2xx, cross-module helpers (json_response, json_with_status, error_json, html_response, live_runners_snapshot, validate_open_path). Routes are wired exactly as before.
- shell.rs — /, /fragment, /check-updates, /search-history, favicons, vendor + dist + asset static routes. Owns the serve_embedded / serve_under / serve_fixed helpers since no one outside this module calls them.
- steps.rs — six step-mutation routes (/toggle, /add-step, /remove-step, /edit-step, /set-priority, /reorder-all) + their payload structs.
- notes.rs — seven file-level routes (GET /note, /note-raw, POST /note, /note/rename, /note/create, /note/mkdir, /note/upload) + the UPLOAD_MAX_BYTES cap.
- items.rs — /create-item (and /api/items alias).
- events.rs — /events SSE stream.
- terminal.rs — /ws/term handler.
- runners.rs — /api/runner/{start,stop,force-stop} + /ws/runner/:key.
- config_surface.rs — GET + POST /configuration, GET /config summary. Named with the _surface suffix to avoid shadowing the crate-root config module.
- openers.rs — /open, /open-folder, /open-external, /open-doc, /recent-screenshot.
- rescan.rs — /rescan.

Each sub-module exposes its handlers as pub(super), so the route wiring in mod.rs::build_router is the only public entry point; the tests use server::start and server::AppState unchanged.

cargo build --workspace + cargo test --workspace green; no behaviour change.
